Added cylinder cock readers.
