+++ CCL.Importer/Proxies/VFX/VFXProxyReplacer.cs
@@ -11,7 +11,7 @@ namespace CCL.Importer.Proxies.VFX
         public VFXProxyReplacer()
         {
             CreateMap<ParticlesPortReadersControllerProxy, ParticlesPortReadersController>().AutoCacheAndMap()
-                .ReplaceGOs()
+                .ReplaceInstancedObjects()
                 .AfterMap(ParticlesPortReadersControllerAfter);
             CreateMap<ParticlesPortReadersControllerProxy.ParticlePortReader, ParticlesPortReadersController.ParticlePortReader>();
             CreateMap<ParticlesPortReadersControllerProxy.ParticlePortReader.PortParticleUpdateDefinition,
